[Hotfix] 2024-12-06 이미지 로드 오류

+++ scripts/before_install.sh
@@ -1,2 +1,4 @@
 #!/bin/bash
-rm -f /home/ubuntu/instagram-server/my.cnf
+if [ -f /home/ubuntu/instagram-server/my.cnf ]; then
+    rm -f /home/ubuntu/instagram-server/my.cnf
+fi
